requirements.txt


aiofiles>=0.5.0
aiohttp>=3.6.2
humanfriendly>=8.2
tqdm>=4.49.0
trafaret>=2.1.0
PyYAML>=5.3.1
spacy>=2.3.2<3.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm==2.2.0

